+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Gamepad/Mouse functions inside InputSystem class
+
 ## [2.4.1] - 2026-07-10
 ### Fixed
 - Only apply rumble effect when using a Gamepad